record-summarize-api/requirements.txt
numpy
pandas
torchaudio
torch
tqdm
fastapi[standard]
uvicorn
pyyaml
fuzzywuzzy
python-Levenshtein
transformers
sentencepiece
soundfile
sqlalchemy
asyncpg
sqlalchemy[asyncio]
pydantic
alembic
dotenv
psycopg2
pydantic_settings
fastcrud
bcrypt
Authlib
pgvector[sqlalchemy]
websockets
fastapi-mail
boto3
botocore
google-auth
itsdangerous
PyJWT[crypto]
asyncio
aioredis
pytest
#ctc_segmentation
google-genai
openai
PyPDF2
python-docx
langchain